Fix and add comments for models
