+++ db/schema.sql
@@ -1,0 +1,3 @@
+DROP DATABASE IF EXISTS rate_my_bootcamp;
+
+CREATE DATABASE rate_my_bootcamp;
